requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
supabase>=2.13.0
python-dotenv>=1.0.1
httpx>=0.28.0
pydantic>=2.10.0
